Move supabase scripts into a legacy folder
