+++ apps/database/src/scripts/reg-types.ts
