+++ apps/database/src/scripts/supabase/shared.ts
